added data attributes to ObjectiveOptions and renamed folder to match jsx
