+++ client/src/components/ChallengeCard/components/ObjectiveOptions/ObjectiveOptions.jsx
@@ -1,8 +1,9 @@
 import { useState } from 'react';
+import { useChallengeContext } from '../../ChallengeContext';
 
 const ObjectivesSelect = (props) => {
 
-    // console.log(props);
+    const [{name}, ] = useChallengeContext();
 
     const [objProgress, setObjProgress] = useState(() => {
         return(props.progress);
@@ -14,7 +15,13 @@ const ObjectivesSelect = (props) => {
     };
 
     return (
-        <select id={`${props.task.replaceAll(' ', '-')} ${props.index}`} value={objProgress} onChange={handleSelect}>
+        <select 
+            id={`${props.task.replaceAll(' ', '-')}`}
+            value={objProgress}
+            data-challenge={name.replaceAll(' ', '-')}
+            data-index={props.index}
+            onChange={handleSelect}
+        >
             <option key='default' disabled>Your Progress</option>
             {[...Array((props.goal + 1))].map((num, i) => {
                 return (<option key={i} value={i} >{i}</option>)
